Fixed some bug and add some example

+++ src/main.v
@@ -35,7 +35,7 @@ fn main() {
 	fp.skip_executable()
 
 	tape_size := fp.int('size', 0, 8192, 'Tape Size')
-	tape_type := fp.string('type', 0, 'u8', 'Tape Type')
+	tape_type := fp.string('type', 0, 'int', 'Tape Type')
 
 	translate := fp.bool('translate', `t`, true, 'New Line Translation (Off)')
 	emit_v := fp.bool('emitv', 0, false, 'Emit V File')
